CAuthentication and Authorization complete

+++ bollywoodhub/app/Http/Controllers/MovieController.php
@@ -8,96 +8,70 @@ use App\Http\Requests\UpdateMovieRequest;
 use Illuminate\Support\Facades\Session;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 
 class MovieController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+    use AuthorizesRequests;
+    // public function __construct()
+    // {
+    //     $this->authorizeResource(Movie::class, 'movie');
+    // }
+
     public function index()
     {
         return view('movies.index', [
             'movies' => Movie::all()
         ]);
-
     }
 
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create()
     {
+        $this->authorize('create', Movie::class);
         return view('movies.create');
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(StoreMovieRequest $request)
     {
+        $this->authorize('create', Movie::class);
         Movie::create($request->validated());
-
         Session::flash('success', 'Movie added successfully');
-        return redirect() -> route('movies.index');
-
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(Movie $movie)
-    {
-        return view('movies.show', compact('movie'));
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(Movie $movie)
-    {
-        return view('movies.edit', compact('movie'));
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(UpdateMovieRequest $request, Movie $movie)
-    {
-        $movie->update($request->validated());
         return redirect()->route('movies.index');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function  trashed()
+    public function show(Movie $movie)
     {
-         return view('movies.trashed',[
-        'movies' => Movie::onlyTrashed()->get()// equivalent to select * from students 
-        ]);
-    }
-    public function trash($id)
-    {
-        Movie::Destroy($id);
-        Session::Flash('success', 'Movie trashed successfully');
-        return redirect() -> route('movies.index');
+        
+        return view('movies.show', compact('movie'));
     }
 
-    public function destroy($id)
+    public function edit(Movie $movie)
     {
-        $movie = Movie::withTrashed() -> where('id', $id) -> first();
-        $movie -> forceDelete();
-        Session::Flash('success', 'Movie deleted successfully');
-        return redirect() -> route('movies.index');
+        $this->authorize('update', $movie);
+        return view('movies.edit', compact('movie'));
     }
 
-    public function restore($id)
+    public function update(UpdateMovieRequest $request, Movie $movie)
     {
-        $movie = Movie::withTrashed() -> where('id', $id) -> first();
-        $movie -> restore();
-        Session::Flash('success', 'Movie restored successfully');
-        return redirect() -> route('movies.trashed');
+        $this->authorize('update', $movie);
+        $movie->update($request->validated());
+        Session::flash('success', 'Movie Updated successfully');
+        return redirect()->route('movies.index');
     }
 
+   
+
+   
+
+    public function destroy(Movie $movie)
+    {
+        
+        $this->authorize('forceDelete', $movie);
+        $movie->forceDelete();
+        Session::flash('success', 'Movie deleted Successfully');
+        return redirect()->route('movies.index');
+    }
+
+    
 }
